D-8: the fuel-cell controller that could not run leaves with its three legacy configs
`advanced_fuel_cell_controller` has no call site anywhere, and it could not have served one.
`ExtendedControllerSimulation.control_chp`, `control_gas_heater` and their siblings read
`ExtendedControllerConfig` as bare class attributes — `ExtendedControllerConfig.chp`,
`.chp_mode`, `.chp_power_states_possible`, `.maximum_autarky` — and that class is a `@dataclass`
with no field defaults, so the class object carries no such attributes and the first control
call raises `AttributeError`. Its test passes because it exercises `_sensor_index` and
`__init__` and never reaches a control method. The component and
`tests/test_advanced_fuel_cell_controller.py` therefore move to `obsolete/components/` and
`obsolete/tests/` together, so the test stays next to the code it tested.

The three configurations go with it, because this controller was their only user in the
repository: `ExtendedControllerConfig` (its configuration), and `GasControllerConfig` and
`CHPControllerConfig` (the hysteresis constants `ExtendedControllerSimulation` reads). They
lived in `hisim/components/configuration.py`, which is a shared file that many live components
still import, so they are moved *out* of it rather than deleted with it: the new
`obsolete/components/configuration_fuel_cell_controller.py` holds all three verbatim under a
docstring saying where they came from and why they left. `configuration.py` keeps everything
else, `GasHeaterConfig` and `PhysicsConfig` included, and the file's remaining classes are
untouched.

Two tables that name the class by string lose their row. `DeviceEnergySpecs.BY_CLASS_NAME` in
`hisim/economics/adapter.py` carried `"ExtendedController": ()` — an explicit "this component
publishes no energy flow the balance draws" — and `tests/test_economics_extraction.py` bound
that key to `hisim.components.advanced_fuel_cell_controller` in the class-to-module table its
contract test walks. A key naming no class in `hisim.components` fails that test by design, so
both go. `tests/test_economics_extraction.py` is green: 93 passed.

Verified: nothing outside `obsolete/` and the roadmap names the module or any of the three
configurations; the energy management system declares no default connections for the class, so
no dynamic output is created or renumbered; no setup builds it, so no golden reference,
scenario JSON or energy-system file changes.
