+++ obsolete/components/advanced_fuel_cell_controller.py
