+++ tests/test_economics_extraction.py
@@ -1297,7 +1297,6 @@ _ENERGY_BALANCE_MODULES = {
     "UtspLpgConnector": "loadprofilegenerator_utsp_connector",
     "ElectricityMeter": "electricity_meter",
     "L2GenericEnergyManagementSystem": "controller_l2_energy_management_system",
-    "ExtendedController": "advanced_fuel_cell_controller",
     "FuelCellController": "controller_l1_fuel_cell",
     "L1Controller": "controller_l1_generic_ev_charge",
     "L1GenericElectrolyzerController": "controller_l1_electrolyzer",
